// sourcecode/BTrak.Packages/snova-profile-module/Profile-module-components/profile-components/src/lib/snovasys-profile-management/models/performance-submission.model.ts
export class PerformanceSubmissionModel {
    performanceId: string;
    configurationId: string;
    performanceDetailsId: string;
    isOpen: boolean;
    isShare: boolean;
    isArchived: boolean;
    pdfUrl: string;
    ofUserId: string;
    formData: string;
    formJson: string;
    submissionFrom: number;
    isCompleted: boolean;
    submittedBy: string;
    submittedByName: string;
    submittedByProfileImage: string;
    createdDateTime: any;
    submittedOn: Date;
    totalCount: number;
    pageSize: number;
    pageNumber: number;
    sortBy: string;
    sortDirectionAsc: boolean;
    searchText: string;
    createdByUserId: string;
    createdByUserName: string;
    createdByUserImage: string;
    closedByUserId: string;
    closedByUserName: string;
    closedByUserImage: string;
    latestModificationOn: Date;
    configurationName: string;
    employeeId: string;
    ofUserName: string;
}